requirements/base.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.0
tablib==0.12.1
django-import-export
django-rest-swagger
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-environ==0.8.1
django-model-utils==4.2.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.1.1
psycopg2==2.8.6
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0